+++ components/Button/vue/index.ts
@@ -2,5 +2,6 @@ export { default } from './Button.vue'
 export {
   VariantClassesTable,
   SizeClassesTable,
+  StaticClasses,
   type ButtonVariants,
 } from '@cypress-design/constants-button'