+++ components/Tooltip/vue/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Tooltip.vue'
+export { type Placement } from '@floating-ui/dom'