+++ components/Button/vue/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @cypress-design/vue-button
 
+## 0.11.12
+
+### Patch Changes
+
+- [#439](https://github.com/cypress-io/cypress-design/pull/439) [`5bddc81`](https://github.com/cypress-io/cypress-design/commit/5bddc8150ae23549c3dca059a79260914323836a) Thanks [@rachelruderman](https://github.com/rachelruderman)! - - Tabs component now requires `aria-controls` prop
+  - Tab `id` is now passed through as an `id` attribute on the tab
+  - Inactive tabs now have `aria-selected=false`
+  - Button component now exports StaticClasses
+
+## 0.11.11
+
+### Patch Changes
+
+- [`04f8d3e`](https://github.com/cypress-io/cypress-design/commit/04f8d3e68b7d222e89897d9ca046f98034ece1a6) Thanks [@elevatebart](https://github.com/elevatebart)! - add bg white to outline-disabled
+
+- Updated dependencies [[`04f8d3e`](https://github.com/cypress-io/cypress-design/commit/04f8d3e68b7d222e89897d9ca046f98034ece1a6)]:
+  - @cypress-design/constants-button@0.1.8
+
+## 0.11.10
+
+### Patch Changes
+
+- [`29703c4`](https://github.com/cypress-io/cypress-design/commit/29703c4d4d8fbb4f742bf57fb7577b41174a6590) Thanks [@elevatebart](https://github.com/elevatebart)! - fix height and width of buttons to match the size of non-square buttons
+
+- Updated dependencies [[`29703c4`](https://github.com/cypress-io/cypress-design/commit/29703c4d4d8fbb4f742bf57fb7577b41174a6590)]:
+  - @cypress-design/constants-button@0.1.7
+
 ## 0.11.9
 
 ### Patch Changes